+++ ledger-wasm/ledger-v8.template.d.ts
@@ -1552,7 +1552,7 @@ export class ZswapChainState {
    * @param tblock - timestamp of a block last batch of updates was applied at
    * @param retentionDuration - number of seconds to retain past Merkle tree roots
    */
-  postBlockUpdate(tblock: Date, retentionDuration?: bigint): ZswapChainState;
+  postBlockUpdate(tblock: Date, retentionDuration: bigint): ZswapChainState;
 
   /**
    * Try to apply an {@link ZswapOffer} to the state, returning the updated state