+++ ledger-wasm/src/zswap_state.rs
@@ -387,18 +387,15 @@ impl ZswapChainState {
     pub fn post_block_update(
         &self,
         tblock: &Date,
-        retention_duration: Option<BigInt>,
+        retention_duration: BigInt,
     ) -> Result<ZswapChainState, JsError> {
         let block_timestamp = Timestamp::from_secs(js_date_to_seconds(tblock));
 
-        let new_inner = match retention_duration {
-            Some(duration_secs) => {
-                let duration = i128::try_from(duration_secs)
-                    .map_err(|_| JsError::new("retention_duration is out of range"))
-                    .map(Duration::from_secs)?;
-                self.0.post_block_update_with(block_timestamp, duration)
-            }
-            None => self.0.post_block_update(block_timestamp),
+        let new_inner = {
+            let duration = i128::try_from(duration_secs)
+                .map_err(|_| JsError::new("retention_duration is out of range"))
+                .map(Duration::from_secs)?;
+            self.0.post_block_update_with(block_timestamp, duration)
         };
 
         Ok(ZswapChainState(new_inner))